mysql_tools/requirements.txt
-e git://github.com/django/django.git@20a91cce04c72bc8c64a1c43b7398edac7b709cc#egg=Django-dev
PIL==1.1.7
MySQL-python==1.2.4
South==0.7.6
humanize==0.3
python-rrdtool==1.4.7
pytz
sqlparse==0.1.6
mmh3==2.0
http://sourceforge.net/projects/pylibpcap/files/pylibpcap/0.6.4/pylibpcap-0.6.4.tar.gz/download
construct==2.06
django-crispy-forms==1.2.3
